Remove ignored files part 1
Didn't properly set .gitignore for use with Android Studio.  As a result, many files that didn't need to be committed (and probably shouldn't be committed), were committed.

In order to fix, have to remove all affected files and this is the easiest way.

All non-ignored files will be back next commit.