Rename camera layout, finalize perspective transform, remove test flashlight functionality, and add rough click interaction
After spending about 8 hours trying to get the flashlight to work while the camera was on, I gave up.  It just doesnt seem to be possible with the current Android SDK without modifiying their source code.

Even though I haven't found it mentioned anywhere, it's probably intentional.  For some reason, even though it was possible in the past, the Android devs don't want users to be able to stream the camera and have flash on at the same time.